[♻️refactor]: Requirement 폴더를 KanbanBoard 폴더 내부로 이동
