+++ src/components/KanbanBoard/Requirement/RequirementControlBtn.tsx
@@ -10,11 +10,11 @@ import { theme, ThemeType } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
 import { copyText, generateSearchPrompt } from '@/utils/gpt';
 
-import GPTGenerateCodeBtn from '../Gpt/GPTGenerateCodeBtn';
-import DeleteHoverBtn from '../Shared/Button/DeleteHoverBtn';
-import HoverIcon from '../Shared/Icons/HoverIcon';
-import SearchIcon from '../Shared/Icons/SearchIcon';
-import DeleteConfirmToast from '../Shared/Toast/DeleteConfirmToast';
+import GPTGenerateCodeBtn from '../../Gpt/GPTGenerateCodeBtn';
+import DeleteHoverBtn from '../../Shared/Button/DeleteHoverBtn';
+import HoverIcon from '../../Shared/Icons/HoverIcon';
+import SearchIcon from '../../Shared/Icons/SearchIcon';
+import DeleteConfirmToast from '../../Shared/Toast/DeleteConfirmToast';
 
 type RequirementControlBtnProps = {
   requirement: RequirementStateType;
