+++ src/components/KanbanBoard/Requirement/RequirementInput.tsx
@@ -6,10 +6,10 @@ import { useRequirement } from '@/hooks/useRequirement';
 import { theme, ThemeType } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
 
-import HoverIcon from '../Shared/Icons/HoverIcon';
-import Textarea from '../Shared/Textarea';
-import EmptyCircleIcon from '../Shared/Icons/EmptyCircleIcon';
-import PlusIcon from '../Shared/Icons/PlusIcon';
+import HoverIcon from '../../Shared/Icons/HoverIcon';
+import Textarea from '../../Shared/Textarea';
+import EmptyCircleIcon from '../../Shared/Icons/EmptyCircleIcon';
+import PlusIcon from '../../Shared/Icons/PlusIcon';
 
 type RequirementInputProps = {
   issueId: string;
